+++ Kosmobook/users/forms.py
@@ -2,6 +2,7 @@ from django import forms
 from django.contrib.auth.forms import UserCreationForm
 from .models import CustomUser
 
+# User Registration
 class UserRegistration(UserCreationForm):
     email = forms.EmailField(required=True)
 
@@ -16,3 +17,22 @@ class UserRegistration(UserCreationForm):
         if commit:
             user.save()
         return user
+
+# Edit Profile   
+class EditProfileForm(forms.ModelForm):
+    class Meta:
+        model = CustomUser
+        fields = [
+            'username',
+            'email',
+            'bio',
+            'profile_picture',
+            'date_of_birth',
+            'location',
+            'specialty',
+            'years_of_experience',
+            'licensed',
+        ]
+        widgets = {
+            'bio': forms.Textarea(attrs={'rows': 3}),
+        }
